move version/index.ts to version/query/get.ts
