+++ src/api/version/query/get.ts
@@ -1,6 +1,6 @@
 import type { QueryFunctionContext } from "@tanstack/react-query";
-import { VersionSchema } from "./schema";
-import { apiFactory } from "../apiFactory";
+import { VersionSchema } from "../schema";
+import { apiFactory } from "../../apiFactory";
 import { useApiKey } from "~/util/store/apiKey";
 import { useQuery } from "@tanstack/react-query";
